feat: added display of likes on the page /movies
