+++ src/components/Movies/Movies.js
@@ -21,6 +21,7 @@ function Movies(props) {
             searchFailed={props.searchFailed}
             saveMovie={props.saveMovie}
             deleteMovie={props.deleteMovie}
+            savedMovies={props.savedMovies}
           />
         )}
       </main>
